refactor: 관련있는 패키지끼리 응집
