+++ src/main/java/chess/domain/movement/policy/NoRestrictionPolicy.java
@@ -1,6 +1,6 @@
 package chess.domain.movement.policy;
 
-import chess.domain.Color;
+import chess.domain.piece.Color;
 
 public class NoRestrictionPolicy implements Policy {
