+++ src/main/java/chess/domain/movement/policy/Policy.java
@@ -1,6 +1,6 @@
 package chess.domain.movement.policy;
 
-import chess.domain.Color;
+import chess.domain.piece.Color;
 
 public interface Policy {
     boolean isSatisfied(Color color, boolean firstMove, boolean existEnemy);
