+++ src/main/java/chess/view/PieceSymbol.java
@@ -1,8 +1,8 @@
 package chess.view;
 
-import chess.domain.Color;
-import chess.domain.Piece;
-import chess.domain.PieceType;
+import chess.domain.piece.Color;
+import chess.domain.piece.Piece;
+import chess.domain.piece.PieceType;
 import java.util.Arrays;
 
 public enum PieceSymbol {
