+++ src/test/java/chess/domain/movement/policy/ColorPolicyTest.java
@@ -2,7 +2,7 @@ package chess.domain.movement.policy;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.Color;
+import chess.domain.piece.Color;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
