+++ src/test/java/chess/domain/movement/policy/FirstMovePolicyTest.java
@@ -3,7 +3,7 @@ package chess.domain.movement.policy;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import chess.domain.Color;
+import chess.domain.piece.Color;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
